+++ docs/changelog.md
@@ -11,6 +11,10 @@ All notable changes to this project will be documented below.
 
 * v0.1dev: **geospatial.analyze.height_percentile**(*dsm, geojson, lower=25, upper=90, label=None*)
 
+#### geospatial.analyze.spectral_index
+
+* v0.1dev: **geospatial.analyze.spectral_index**(*img, geojson, percentiles=None, label=None*)
+
 #### geospatial.points2roi_circle
 
 * v0.1dev: rois = **geospatial.points2roi_circle**(*img, geojson, radius*)
@@ -21,7 +25,7 @@ All notable changes to this project will be documented below.
 
 #### geospatial.center_grid_rois
 
-* v0.1dev: rois = **geospatial.center_grid_rois**(*img, viewer, radius=10*)
+* v0.1dev: rois = **geospatial.center_grid_rois**(*img, viewer, radius=10, layername="Shapes"*)
 
 #### geospatial.napari_grid
 
@@ -29,7 +33,7 @@ All notable changes to this project will be documented below.
 
 #### geospatial.napari_polygon_grid
 
-* v0.1dev: **geospatial.napari_polygon_grid**(*viewer*)
+* v0.1dev: **geospatial.napari_polygon_grid**(*viewer, layername="Shapes"*)
 
 #### geospatial.read_geotif
 
@@ -58,4 +62,4 @@ All notable changes to this project will be documented below.
 
 #### geospatial.shapes_to_geojson
 
-* v0.1dev: coord = **geospatial.shapes_to_geojson**(*img, viewer, out_path*)
+* v0.1dev: coord = **geospatial.shapes_to_geojson**(*img, viewer, out_path, layername="Shapes"*)